refactor: reorganize component imports for clarity and consistency
